swenet/requirements.txt
numpy
pyyaml
pillow